requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
python-dotenv
slowapi
logfire
pyyaml
requests
